rip bozo :100: packwatch

+++ RSDKv5/Math.cpp
@@ -78,7 +78,7 @@ void CalculateTrigAngles()
     for (int y = 0; y < 0x100; ++y) {
         uint8 *arcTan = (uint8 *)&atanVal256[y];
         for (int x = 0; x < 0x100; ++x) {
-            *arcTan = (uint8)(atan2(y, x) * 40.743664);
+            *arcTan = (int)(float)((float)atan2((float)y, x) * 40.743664);
             arcTan += 0x100;
         }
     }
